expected output for xpath_context input data
This is the expected data for the input data added in
w3c/its-2.0-testsuite-inputdata#5. The purpose of the test is to check
that position and size are set for the XPath contexts of relative
selectors. This is done using `idValue`, which is an XPath that returns
a string, making the test simpler. XPath functions `pos()` and `last()`
are evaluated and concatenated in the `idValue` values.